fix: users can create coffees, modify/delete their own coffees

+++ src/coffees/coffees.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, InternalServerErrorException, Param, Patch, Post, Query, Req, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { PaginationQueryDto } from '../common/dto/pagination-query.dto/pagination-query.dto';
 import { CoffeesService } from './coffees.service';
 import { CreateCoffeeDto } from './dto/create-coffee.dto';
@@ -7,14 +7,17 @@ import { Protocol } from '../common/decorators/protocol.decorator';
 import { ApiForbiddenResponse, ApiTags } from '@nestjs/swagger';
 import { Roles } from '../common/decorators/roles.decorator';
 import { Role } from '../users/entities/user.entity';
-
+import { UsersService } from 'src/users/users.service';
 
 @ApiTags('coffees')
 @Controller({ path: 'coffees', version: '1' })
 export class CoffeesController {
-    constructor(private readonly coffeeService: CoffeesService) {}
+    constructor(
+        private readonly coffeeService: CoffeesService, 
+        private readonly usersService: UsersService,
+    ) {}
 
-    @Get('flavors')
+    @Get('')
     @Roles([Role.Admin, Role.User])
     @ApiForbiddenResponse({ description: 'Forbidden.' })
     async findAll(@Protocol('https') protocol: string, @Query() paginationQuery: PaginationQueryDto) {
@@ -23,26 +26,63 @@ export class CoffeesController {
 
     @Get(':id')
     @Roles([Role.Admin, Role.User])
-    findOne(@Param('id') id: string) {
+    async findOne(@Param('id') id: string) {
         const coffee = this.coffeeService.findOne(id);
         return coffee;
     }
 
     @Post()
-    @Roles([Role.Admin])
-    create(@Body() createCoffeeDto: CreateCoffeeDto) {
-        return this.coffeeService.create(createCoffeeDto);
+    @Roles([Role.Admin, Role.User])
+    async create(@Body() createCoffeeDto: CreateCoffeeDto, @Req() request: any) {
+
+        const username = request.user.username;
+        const inventor = await this.usersService.findOne(username);
+
+        return this.coffeeService.create(createCoffeeDto, inventor);
     }
 
     @Patch(':id')
-    @Roles([Role.Admin])
-    update(@Param('id') id: string, @Body() updateCoffeeDto: UpdateCoffeeDto) {
-        return this.coffeeService.update(id, updateCoffeeDto);
+    @Roles([Role.Admin, Role.User])
+    async update(@Param('id') id: string, @Body() updateCoffeeDto: UpdateCoffeeDto, @Req() request: any) {
+
+        const username = request.user.username;
+        const user = await this.usersService.findOne(username);
+        
+        const coffee = await this.coffeeService.findOne(id);
+
+        if (user.role == Role.Admin) {
+            return await this.coffeeService.update(id, updateCoffeeDto);
+        } else if (user.role == Role.User) {
+            if (coffee.inventor.id == user.id) {
+                return this.coffeeService.update(id, updateCoffeeDto);
+            } else {
+                throw new UnauthorizedException("User cannot update others' coffees");
+            }
+        } else {
+            throw new InternalServerErrorException("Unsupported user role");
+        }
     }
 
     @Delete(':id')
-    @Roles([Role.Admin])
-    remove(@Param('id') id:string) {
-        return this.coffeeService.remove(id);
+    @Roles([Role.Admin, Role.User])
+    async remove(@Param('id') id:string, @Req() request: any) {
+
+        const username = request.user.username;
+        const user = await this.usersService.findOne(username);
+
+        const coffee = await this.coffeeService.findOne(id);
+
+        if (user.role == Role.Admin) {
+            return this.coffeeService.remove(id);
+        } else if (user.role == Role.User) {
+            if (coffee.inventor.id == user.id) {
+                return this.coffeeService.remove(id);
+            } else {
+                throw new UnauthorizedException("User cannot delete others' coffees");
+            }
+        } else {
+            throw new InternalServerErrorException("Unsupported user role");
+        }
     }
+
 }
